requirements.txt
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
keeper-secrets-manager-core==16.6.6
MarkupSafe==3.0.2
pycparser==2.22
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.32.3